fix: move decision helper to service, H1 titles
